requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==3.1.0
distlib==0.3.6
Django==4.1.5
django-cleanup==7.0.0
django-environ==0.10.0
django-filter==22.1
django-mixins==0.0.10
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
filelock==3.9.0
idna==3.4
Pillow==9.4.0
platformdirs==3.1.1
psycopg2==2.9.5
pytz==2022.7.1
requests==2.28.2
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.15
utils==1.0.1
virtualenv==20.21.0
